onto-agent-server/pyproject.toml
[project]
name = "onto-agent-server"
version = "0.1.0"
description = "OntoAgent Backend API Server"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyhumps>=3.8.0",
]

[tool.uv]
dev-dependencies = []
